tui: Adapt to u64 API
Update call sites for u64 addresses, sizes and progress.

+++ tui/src/cli/commands/device/download.rs
@@ -71,7 +71,7 @@ impl DeviceCommand for DownloadArgs {
         info!("Downloading to partition '{}'...", part.name);
 
         if let Err(e) =
-            dev.write_partition(&part.name, file_size as usize, &mut reader, &mut progress_callback)
+            dev.write_partition(&part.name, file_size, &mut reader, &mut progress_callback)
         {
             pb.abandon("Download failed!");
             Err(e)?;
